权限控制初步完成

+++ src/main/java/cn/jihangyu/glowworm/activity/controller/ActivityController.java
@@ -2,11 +2,12 @@ package cn.jihangyu.glowworm.activity.controller;
 
 import cn.jihangyu.glowworm.activity.entity.Activity;
 import cn.jihangyu.glowworm.activity.service.ActivityService;
-import cn.jihangyu.glowworm.common.constants.Constants;
+import cn.jihangyu.glowworm.common.base.BaseController;
 import cn.jihangyu.glowworm.common.enums.ResultEnum;
 import cn.jihangyu.glowworm.common.execption.GlowwormExecption;
 import cn.jihangyu.glowworm.common.resp.ApiResult;
 import cn.jihangyu.glowworm.common.utils.ResultUtil;
+import cn.jihangyu.glowworm.user.entity.UserElement;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,9 +16,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartException;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.Date;
 import java.util.List;
 
 /**
@@ -29,7 +28,7 @@ import java.util.List;
  */
 @RestController
 @RequestMapping("activity")
-public class ActivityController {
+public class ActivityController extends BaseController{
 
     @Autowired
     @Qualifier("activityServiceImpl")
@@ -46,8 +45,15 @@ public class ActivityController {
     @ApiImplicitParam(name = "activity", value = "活动详细实体activity", required = true, dataType = "Activity")
     @RequestMapping(value = "/addActivity", method = RequestMethod.POST)
     public ApiResult addActivity(@RequestBody Activity activity) throws Exception {
-        Activity activity1 = activityService.addActivity(activity);
-        return ResultUtil.success(activity1);
+        Activity activity1=null;
+        UserElement ue=getCurrentUser();
+        if("admin".equals(ue.getRole())){
+            activity1= activityService.addActivity(activity);//管理员才可以创建
+            return ResultUtil.success(activity1);
+        }else {
+            throw new GlowwormExecption(ResultEnum.NO_AUTHORITY);
+        }
+
     }
 
     @ApiOperation(value = "修改活动", notes = "根据活动对象修改用户")
@@ -61,8 +67,14 @@ public class ActivityController {
     @ApiOperation(value = "根据id删除活动", notes = "根据id删除活动")
     @RequestMapping(value = "/deleteActivity/{id}", method = RequestMethod.GET)
     public ApiResult deleteActivity(@PathVariable Integer id) throws Exception {
-        activityService.deleteActivityById(id);
-        return ResultUtil.success("删除成功");
+        UserElement ue=getCurrentUser();
+        if("admin".equals(ue.getRole())){
+            activityService.deleteActivityById(id);
+            return ResultUtil.success("删除成功");
+        }else {
+            throw new GlowwormExecption(ResultEnum.NO_AUTHORITY);
+        }
+
     }
 
     @ApiOperation(value = "查找不同类型的活动", notes = "{1：未进行，2：正在进行，3：已结束，0:所有)")
